added flags + prepared files for bonuses

+++ bonuses/win_handler_bonus.c
@@ -6,7 +6,7 @@
 /*   By: alarose <alarose@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 12:05:30 by alarose           #+#    #+#             */
-/*   Updated: 2024/06/27 14:38:52 by alarose          ###   ########.fr       */
+/*   Updated: 2024/06/27 10:38:15 by alarose          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int	handle_input(int keysym, t_data *data)
 	if (data->map.map_layout[data->img[PLAYER].y][data->img[PLAYER].x] == 'P')
 		data->map.move_out_i_img = BG;
 	i = 0;
-	while ((long unsigned)i < (sizeof(f_table) / sizeof(f_table[0])))
+	while (i < (sizeof(f_table) / sizeof(f_table[0])))
 	{
 		if (f_table[i].keycode == keysym)
 		{
